Migrate decompose_sfhpop_into_bulge_disk_knots from photometry_lc_interp.py module into disk_bulge_kernels.py module
